refactor: 게임 진행 부분 새로운 클래스로 이동
