+++ src/main/java/racingcar/Application.java
@@ -9,32 +9,9 @@ public class Application {
         List<Car> cars = InputHandler.inputCarNames();
         int tryCount = InputHandler.inputTryCount();
 
-        gameStart(cars, tryCount);
+        GameHandler.gameStart(cars, tryCount);
 
         printWinner(cars);
-    }
-
-    private static void gameStart(List<Car> cars, int tryCount) {
-        System.out.println("실행 결과");
-
-        for (int i = 0; i < tryCount; i++) {
-            for (Car car : cars) {
-                car.move();
-            }
-
-            printProgress(cars);
-        }
-    }
-
-    private static void printProgress(List<Car> cars) {
-        for (Car car : cars) {
-            System.out.print(car.getName() + " : ");
-            for (int i = 0; i < car.getLocation(); i++) {
-                System.out.print("-");
-            }
-            System.out.println();
-        }
-        System.out.println();
     }
 
     private static void printWinner(List<Car> cars) {
